pyproject.toml
[project]
name = "design-intelligence"
version = "0.1.0"
description = "Agent-based platform for expert-level UX/UI design feedback, style guides, and design specifications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.40.0",
    "typer>=0.12.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "pillow>=10.3.0",
    "playwright>=1.44.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0.0",
    "httpx>=0.27.0",
    "rich>=13.7.0",
    "markdown>=3.6",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
vector = [
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
design-intel = "src.cli:app"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
